+++ CHANGELOG.md
@@ -1,5 +1,14 @@
 # CHANGELOG
 
+#### Changed
+- The time taken by the participant is now stored as a property of the `Trial` object
+  (`Trial.time_taken`).
+- By default, dynamic updating of progress bar and bonus display now only occurs
+  for Unity pages. This makes the logs cleaner for standard PsyNet pages.
+
+#### Fixed
+- Updated `get_template` to remove use of deprecated function `read_text`.
+
 # [5.0.1] Released on 2021-11-10
 
 #### Fixed